Fix some compile errors with apple clang
